+++ chunkflow/operators/base.py
