+++ chunkflow/operators/cloud_watch.py
@@ -1,6 +1,6 @@
 from chunkflow.aws.cloud_watch import CloudWatch
 
-from .operator_base import OperatorBase
+from .base import OperatorBase
 
 
 class CloudWatchOperator(OperatorBase):
