+++ chunkflow/operators/crop_margin.py
@@ -1,5 +1,5 @@
 from chunkflow.chunk import Chunk 
-from .operator_base import OperatorBase
+from .base import OperatorBase
 
 
 class CropMarginOperator(OperatorBase):
