+++ chunkflow/operators/patch_inference_engine/base.py
@@ -1,15 +1,21 @@
 import numpy as np
+from .patch_mask import PatchMask
+
 
 class PatchInferenceEngine(object):
-    def __init__(self):
-        pass
+    def __init__(self, patch_size, patch_overlap):
+        self.patch_size = patch_size
+        self.patch_overlap = patch_overlap
+
+        # prepare patch mask
+        self.mask = PatchMask(patch_size, patch_overlap)
 
     def __call__(self, patch):
         """
         this method should be inherited for real implementation
         """
         return NotImplementedError()
-
+    
     def _reshape_patch(self, patch):
         """patch should be a 5d np array
         """
